requirements.txt
pygame~=2.1.2
neat-python
pickle